Move perfect map to the main implementation
